rename scribe.blockchain -> hub.scribe
