--- hub/blockchain/__init__.py
@@ -1,1 +1,0 @@
-from hub.blockchain.network import LBCTestNet, LBCRegTest, LBCMainNet
