+++ hub/scribe/daemon.py
